ref: move types into types folder
