+++ src/types/express.locals.ts
@@ -3,8 +3,10 @@ import { LeanDocument } from 'mongoose'
 
 import { ProcessedFieldResponse } from '../app/modules/submission/submission.types'
 
-import { IPopulatedForm } from './form'
+import { BasicField } from './field'
+import { IEmailFormSchema, IPopulatedForm } from './form'
 import { SpcpSession } from './spcp'
+import { ISubmissionSchema } from './submission'
 
 export type WithForm<T> = T & {
   form: IPopulatedForm
@@ -37,3 +39,54 @@ export type SpcpLocals =
     }
   | { uinFin: string; userInfo: string }
   | { [key: string]: never } // empty object
+
+export interface EmailAutoReplyField {
+  question: string
+  answerTemplate: string[]
+}
+
+export interface EmailJsonField {
+  question: string
+  answer: string
+}
+
+export interface EmailFormField {
+  question: string
+  answer: string
+  fieldType: BasicField
+  answerTemplate: string[]
+}
+
+export interface EmailData {
+  autoReplyData: EmailAutoReplyField[]
+  jsonData: EmailJsonField[]
+  formData: EmailFormField[]
+}
+
+export interface EmailDataForOneField {
+  autoReplyData?: EmailAutoReplyField
+  jsonData?: EmailJsonField
+  formData: EmailFormField
+}
+
+export type WithSubmission<T> = T & { submission: ISubmissionSchema }
+
+export interface IAttachmentInfo {
+  filename: string
+  content: Buffer
+  fieldId: string
+}
+
+export type WithAttachments<T> = T & { attachments: IAttachmentInfo[] }
+
+export type WithEmailData<T> = T & EmailData
+
+export type WithEmailForm<T> = T & { form: IEmailFormSchema }
+
+export type WithEmailModeMetadata<T> = WithEmailData<T> &
+  WithAttachments<T> &
+  WithEmailForm<T>
+
+export type WithAutoReplyEmailData<T> =
+  | (WithForm<T> & WithSubmission<T>)
+  | WithEmailModeMetadata<T>
